DB-15 Controlador NFC

+++ backend/routers/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const { body, validationResult } = require('express-validator');
 const jwt = require('jsonwebtoken');
-const bcrypt = require('bcrypt');
+const bcrypt = require('bcryptjs');
 const verifyToken = require('../middlewares/authMiddleware');
 const User = require('../models/usuario');
 const router = express.Router();
@@ -11,17 +11,17 @@ router.post('/register', [
   body('username').isLength({ min: 3 }).withMessage('El nombre de usuario debe tener al menos 3 caracteres'),
   body('password').isLength({ min: 6 }).withMessage('La contraseña debe tener al menos 6 caracteres'),
   body('role').isIn(['admin', 'employee']).withMessage('El rol no es válido'),
-  // Aquí puedes agregar la validación para area_id si es relevante
-  // body('area_id').optional().isInt().withMessage('El ID de área debe ser un número entero')
+  body('area').isIn(['Informatica', 'Contabilidad', 'Administración']).withMessage('Area no asignada'),
+  body('codigoNfc').notEmpty().withMessage('El código NFC es obligatorio') 
 ], async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array() });
   }
 
-  const { username, password, role, area_id } = req.body;
+  const { username, password, role, area, codigoNfc } = req.body; 
 
-  console.log(`Creando usuario con el rol: ${role}`); // Agrega esta línea para depuración
+  console.log(`Creando usuario con el rol: ${role}, y código NFC: ${codigoNfc}`); 
 
   try {
     const existingUser = await User.findOne({ where: { username } });
@@ -29,26 +29,28 @@ router.post('/register', [
       return res.status(400).json({ error: 'El nombre de usuario ya está en uso' });
     }
 
-    // Crea el nuevo usuario sin hashear la contraseña
+    // Crea el nuevo usuario incluyendo el codigoNfc
     const newUser = await User.create({
       username,
-      password, // Contraseña en texto plano
+      password, 
       role,
-      area_id // Incluye area_id si es necesario
+      area,
+      codigoNfc 
     });
 
-    const token = jwt.sign({ id: newUser.id, role: newUser.role }, process.env.JWT_SECRET, { expiresIn: '1h' });
+    const token = jwt.sign({ id: newUser.id, role: newUser.role, area: newUser.area }, process.env.JWT_SECRET, { expiresIn: '1h' });
 
     res.status(201).json({ message: 'Usuario creado exitosamente', token, role: newUser.role });
   } catch (error) {
-    console.error('Error al crear usuario:', error); // Mejora la depuración
+    console.error('Error al crear usuario:', error); 
     res.status(500).json({ error: 'Error del servidor' });
   }
 });
 
 const verifyRole = (roles) => {
   return (req, res, next) => {
-    const { role } = req.user; // req.user viene del middleware de autenticación JWT
+    // req.user viene del middleware de autenticación JWT
+    const { role } = req.user; 
 
     if (!roles.includes(role)) {
       return res.status(403).json({ error: 'No tienes permiso para acceder a esta ruta' });
@@ -72,7 +74,7 @@ router.get('/dashboard', verifyToken, verifyRole(['admin', 'employee']), (req, r
 router.get('/users', async (req, res) => {
   try {
     const users = await User.findAll({
-      attributes: ['id', 'username', 'role', 'area_id'] // Incluye 'area_id'
+      attributes: ['id', 'username', 'role', 'area', 'codigoNfc'] 
     });
     res.json(users);
   } catch (err) {
@@ -81,26 +83,24 @@ router.get('/users', async (req, res) => {
   }
 });
 
-
-//eliminar un usuario por ID
+// Eliminar un usuario por su ID
 router.delete('/users/:id', async (req, res) => {
-  const { id } = req.params;
-
   try {
-    const user = await User.findByPk(id);
-    if (!user) {
-      return res.status(404).json({ error: 'Usuario no encontrado' });
+    const userId = req.params.id;
+    const deletedUser = await User.destroy({ where: { id: userId } });
+    
+    if (deletedUser) {
+      res.status(200).json({ message: 'Usuario eliminado correctamente' });
+    } else {
+      res.status(404).json({ error: 'Usuario no encontrado' });
     }
-    await User.destroy({ where: { id } });
-    res.json({ message: 'Usuario eliminado correctamente' });
-
-  } catch (err) {
-    console.error('Error al eliminar el usuario:', err);
+  } catch (error) {
+    console.error('Error al eliminar el usuario:', error);
     res.status(500).json({ error: 'Error del servidor' });
   }
 });
 
-//obtener un usuario por ID
+// Obtener un usuario por ID
 router.get('/users/:id', async (req, res) => {
   const { id } = req.params;
 
@@ -116,21 +116,69 @@ router.get('/users/:id', async (req, res) => {
   }
 });
 
+// Login de usuario
 router.post('/login', async (req, res) => {
   const { username, password } = req.body;
 
   try {
     const user = await User.findOne({ where: { username } });
+    if (!user || user.password !== password) {
+      return res.status(400).json({ error: 'Credenciales incorrectas' });
+    }
 
-    if (!user) return res.status(400).json({ error: 'Usuario no encontrado' });
-
-    // Comparar la contraseña proporcionada directamente
-    if (user.password !== password) return res.status(400).json({ error: 'Contraseña incorrecta' });
-
+    // Generar el token de autenticación
     const token = jwt.sign({ id: user.id, role: user.role }, process.env.JWT_SECRET, { expiresIn: '1h' });
 
-    res.json({ token, role: user.role });
+    // Enviar la respuesta con el ID del usuario
+    res.json({
+      token,
+      userId: user.id, 
+      role: user.role,
+    });
   } catch (error) {
+    console.error('Error en la solicitud de inicio de sesión:', error);
+    res.status(500).json({ error: 'Error en el servidor' });
+  }
+});
+
+// Actualizar un usuario por su ID
+router.put('/users/:id', async (req, res) => {
+  try {
+    const userId = req.params.id;
+    const { username, password, role, area, codigoNfc } = req.body; 
+
+    // Encuentra al usuario y actualiza sus datos, incluido el código NFC
+    const updatedUser = await User.update(
+      { username, password, role, area, codigoNfc },  
+      { where: { id: userId } }
+    );
+
+    if (updatedUser) {
+      res.status(200).json({ message: 'Usuario actualizado correctamente' });
+    } else {
+      res.status(404).json({ error: 'Usuario no encontrado' });
+    }
+  } catch (error) {
+    console.error('Error al actualizar el usuario:', error);
+    res.status(500).json({ error: 'Error del servidor' });
+  }
+
+  
+});
+// Buscar un usuario por su código NFC
+router.get('/users/nfc/:codigoNfc', async (req, res) => {
+  const { codigoNfc } = req.params;
+
+  try {
+    const user = await User.findOne({ where: { codigoNfc } });
+    
+    if (user) {
+      res.json({ exists: true });
+    } else {
+      res.json({ exists: false });
+    }
+  } catch (err) {
+    console.error('Error al buscar usuario por código NFC:', err);
     res.status(500).json({ error: 'Error del servidor' });
   }
 });
